% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-ld.R
\name{compute_ld_set}
\alias{compute_ld_set}
\title{Compute LD among all pairs of a set of variants}
\usage{
compute_ld_set(
  gds,
  variant_include,
  ld_methods = "r2",
  sample_include = NULL,
  n_threads = 1L
)
}
\arguments{
\item{gds}{A SeqArray GDS file}

\item{variant_include}{THe variant.ids for which to calculate ld}

\item{ld_methods}{Character vector of methods to use to calculate LD. Can be any of ``"r2"\verb{, }"r"`, `"dprime"`. See details.}

\item{sample_include}{A vector of sample.ids to use for LD calculation.}

\item{n_threads}{The number of threads to use for LD calculation.}
}
\value{
A data frame with the following columns:
\itemize{
\item variant.id.1: the first variant id in the pair
\item variant.id.2: the second variant id in the pair
\item One column for each method specified giving the LD btween \code{variant.id.1} and \code{variant.id.2} using the specified method:
\itemize{
\item if \code{methods} contains \code{"composite"}: \code{ld_composite}
\item if \code{methods} contains \code{dprime}: \code{"ld_dprime"}
\item if \code{methods} contains \code{"corr"}: \code{ld_corr}
\item if \code{methods} contains \code{"r"}: \code{ld_r}
}
}

Each pair of variants only has one record in the data frame.
}
\description{
Compute LD among all pairs of a set of variants
}
\details{
This function computes the LD between all pairs in a set of variants using \code{snpgdsLDMat} using the specified methods.
The \code{r2} method runs the \code{SNPRelate} \code{corr} method and squares the result.
The \code{r} method runs the \code{SNPRelate} \code{corr} method but does not square the result.
The \code{dprime} method runs the \code{SNPRelate} \code{dprime} method.
}
\examples{

gds <- SeqArray::seqOpen(system.file("extdata", "1KG_phase3_subset.gds", package="LDcompute"))

# Different methods
compute_ld_set(gds, c(1, 2, 3))
compute_ld_set(gds, c(1, 2, 3), ld_methods = "dprime")
compute_ld_set(gds, c(1, 2, 3), ld_methods = c("r2", "dprime"))

# Different sample set
sample_include <- SeqArray::seqGetData(gds, "sample.id")[1:500]
compute_ld_set(gds, c(1, 2, 3), ld_methods = "r2", sample_include = sample_include)

SeqArray::seqClose(gds)

}
\seealso{
\itemize{
\item \code{\link[=compute_ld_pair]{compute_ld_pair()}} computes LD between a pair of variants
\item \code{\link[=compute_ld_set]{compute_ld_set()}} computes LD between all pairs of a set of variants
\item \code{\link[=compute_ld_index]{compute_ld_index()}} computes LD between one variant and a set of other variants
}
}
