% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-ld.R
\name{compute_ld}
\alias{compute_ld}
\title{Compute LD among variants}
\usage{
compute_ld(
  gds,
  variant_include_1,
  variant_include_2 = NULL,
  methods = c("composite")
)
}
\arguments{
\item{gds}{A SeqArray GDS file}

\item{variant_include_1}{A vector of variant ids. See details.}

\item{variant_include_2}{A vector of variant.ids. See details.}

\item{methods}{A vector of methods to use to calculate LD. See details.}
}
\description{
Compute LD among variants
}
\details{
This function is primarily a wrapper around the SNPRelate snpgdsLDMat function.
It allows LD calculations between a pair of variants and returns results in a consistent format.
}
